requirements.txt
flask
flask-cors
python-dotenv
google-genai
openai-whisper
gTTS
torch
numpy
